--- js/functions.js
@@ -1,24 +1,0 @@
-// Проверка длины строки
-const stringLength = (string, length) => string.length <= length;
-
-// Проверка, является ли строка палиндромом
-const isPalindrome = (string) => {
-  string = string.toLowerCase().replaceAll(' ', ''); // Преобразует регистр и удаляет пробелы
-  return string === string.split('').reverse().join('');
-};
-
-// Извлечение цифр из строки
-const extractDigits = (string) => parseInt(String(string).replace(/[^0-9]/g, ''), 10);
-
-// Заполнение строки
-const fillString = (string, minLength, filler) => {
-  while (string.length < minLength) {
-    string = filler.slice(0, minLength - string.length) + string;
-  }
-  return string;
-};
-
-stringLength('Это строка', 12);
-isPalindrome('Не гни папин ген');
-extractDigits('Проблема 2000');
-fillString('Это строка', 16, 'абвгд');
